Fix assimp vertices not being transformed.
WARNING : Transformation is put directly into assimp vertices, and it does not reset the node transformation.

+++ process/process_compute_min_max.cpp
@@ -8,7 +8,7 @@ void process_compute_min_max(process_t& process)
 	process.min.set(std::numeric_limits<float>::max());
 	process.max.set(-std::numeric_limits<float>::max());
 
-	for (uint32 mesh_it = 0; mesh_it < process.scene->mNumMeshes; mesh_it++)
+	for (size_t mesh_it = 0; mesh_it < process.scene->mNumMeshes; mesh_it++)
 	{
 		auto mesh = process.scene->mMeshes[mesh_it];
 		if (mesh->mPrimitiveTypes & (aiPrimitiveType_LINE | aiPrimitiveType_POINT)) continue;
@@ -16,9 +16,9 @@ void process_compute_min_max(process_t& process)
 		if (is_tag(mesh->mName.C_Str()))
 		{
 			continue;
-		}
+		}		
 
-		for (uint32 vert_it = 0; vert_it < mesh->mNumVertices; vert_it++)
+		for (size_t vert_it = 0; vert_it < mesh->mNumVertices; vert_it++)
 		{
 			auto& v = mesh->mVertices[vert_it];
 			process.min.min(v3(v.x, v.y, v.z));
